Work with reflection and the object editor dialog
I should make more commits (everyday I work, not when I finish stuff)

+++ 3DRadSpace/3DRadSpaceDll_Tests/LinearAlgebraTests.cpp
@@ -31,14 +31,14 @@ namespace LinearAlgebraTests
 	TEST(LinearTransformation, MatrixTests)
 	{
 		Engine3DRadSpace::Matrix m = Engine3DRadSpace::Matrix::CreateScale(5);
-		m *= Engine3DRadSpace::Matrix::CreateRotationY(std::numbers::pi / 2);
+		m *= Engine3DRadSpace::Matrix::CreateRotationY((float)(std::numbers::pi / 2));
 		m *= Engine3DRadSpace::Matrix::CreateTranslation({ 1,2,3 });
 
 		Engine3DRadSpace::Matrix expected_matrix(
-			-2.18556949e-07, 0, -5, 0,
-			0, 5, 0, 0,
-			5, 0, -2.18556949e-07, 0,
-			1, 2, 3, 1
+			-2.18556949e-07f, 0.f, -5.f, 0.f,
+			0.f, 5.f, 0.f, 0.f,
+			5.f, 0.f, -2.18556949e-07f, 0.f,
+			1.f, 2.f, 3.f, 1.f
 		);
 
 		EXPECT_EQ(m, expected_matrix);
